1. 調整 Burgers 命名空間下的類別移動至 Models.Burgers 命名空間下，並調整相關參考的程式碼。 2. 格式化與清理 Burgers.cs 文件並增加 XML 註解。 3. 格式化與清理 NewYorkChickenBurger.cs 文件並增加 XML 註解。 4. 格式化與清理 NewYorkPorkBurger.cs 文件並增加 XML 註解。 5. 格式化與清理 TaipeiChickenBurger.cs 文件並增加 XML 註解。 6. 格式化與清理 TaipeiPorkBurger.cs 文件並增加 XML 註解。
