--- NineYi.BurgerShop/Burgers/NewYorkChickenBurger.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NineYi.BurgerShop.Burgers
-{
-    /// <summary>
-    /// 紐約香雞堡
-    /// </summary>
-    public class NewYorkChickenBurger : Burger
-    {
-        public NewYorkChickenBurger()
-        {
-            this.Name = "NewYork Chicken Burger";
-        }
-    }
-}
