--- NineYi.BurgerShop/Burgers/NewYorkPorkBurger.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NineYi.BurgerShop.Burgers
-{
-    /// <summary>
-    /// 紐約豬肉堡
-    /// </summary>
-    public class NewYorkPorkBurger : Burger
-    {
-        public NewYorkPorkBurger()
-        {
-            this.Name = "NewYork Pork Burger";
-        }
-    }
-}
