--- NineYi.BurgerShop/Burgers/TaipeiChickenBurger.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NineYi.BurgerShop.Burgers
-{
-    /// <summary>
-    /// 台北香雞堡
-    /// </summary>
-    public class TaipeiChickenBurger : Burger
-    {
-        public TaipeiChickenBurger()
-        {
-            this.Name = "Taipei Chicken Burger";
-        }
-    }
-}
