--- NineYi.BurgerShop/Burgers/TaipeiPorkBurger.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NineYi.BurgerShop.Burgers
-{
-    /// <summary>
-    /// 台北豬肉堡
-    /// </summary>
-    public class TaipeiPorkBurger : Burger
-    {
-        public TaipeiPorkBurger()
-        {
-            this.Name = "Taipei Pork Burger";
-        }
-    }
-}
