+++ NineYi.BurgerShop/Models/Burgers/NewYorkChickenBurger.cs
@@ -1,0 +1,17 @@
+﻿namespace NineYi.BurgerShop.Models.Burgers
+{
+    /// <summary>
+    /// 紐約香雞堡。
+    /// </summary>
+    /// <seealso cref="NineYi.BurgerShop.Models.Burgers.Burger" />
+    public class NewYorkChickenBurger : Burger
+    {
+        /// <summary>
+        /// 初始化 <see cref="NewYorkChickenBurger"/> 類別新的執行個體。
+        /// </summary>
+        public NewYorkChickenBurger()
+        {
+            this.Name = "NewYork Chicken Burger";
+        }
+    }
+}
