+++ NineYi.BurgerShop/Models/Burgers/TaipeiChickenBurger.cs
@@ -1,0 +1,17 @@
+﻿namespace NineYi.BurgerShop.Models.Burgers
+{
+    /// <summary>
+    /// 台北香雞堡。
+    /// </summary>
+    /// <seealso cref="NineYi.BurgerShop.Models.Burgers.Burger" />
+    public class TaipeiChickenBurger : Burger
+    {
+        /// <summary>
+        /// 初始化 <see cref="TaipeiChickenBurger"/> 類別新的執行個體。
+        /// </summary>
+        public TaipeiChickenBurger()
+        {
+            this.Name = "Taipei Chicken Burger";
+        }
+    }
+}
