+++ NineYi.BurgerShop/Models/Burgers/TaipeiPorkBurger.cs
@@ -1,0 +1,17 @@
+﻿namespace NineYi.BurgerShop.Models.Burgers
+{
+    /// <summary>
+    /// 台北豬肉堡。
+    /// </summary>
+    /// <seealso cref="NineYi.BurgerShop.Models.Burgers.Burger" />
+    public class TaipeiPorkBurger : Burger
+    {
+        /// <summary>
+        /// 初始化 <see cref="TaipeiPorkBurger"/> 類別新的執行個體。
+        /// </summary>
+        public TaipeiPorkBurger()
+        {
+            this.Name = "Taipei Pork Burger";
+        }
+    }
+}
